deriveNativeScriptHash: update for stax
